% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_viz.R
\name{clique_categorical_palette}
\alias{clique_categorical_palette}
\title{The cliqueR default categorical colour palette}
\usage{
clique_categorical_palette()
}
\value{
A character vector of 8 hex colours.
}
\description{
Eight colourblind-safe hues (validated for colour-vision-deficiency separation
on a light surface), assigned to categories in order.
}
