% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliqueR.R
\name{cliqueR.maximal}
\alias{cliqueR.maximal}
\title{Maximal clique enumeration}
\usage{
cliqueR.maximal(filename, least = 3, most = -1, profile = 0)
}
\arguments{
\item{filename}{Character path to an edgelist graph file.}

\item{least}{Integer. Lower bound of clique size to include. (a value less than 3 will simply
return every edge in the graph).}

\item{most}{Integer. Upper bound of clique size to include.}

\item{profile}{Integer. Option to return a clique profile instead of all cliques.}
}
\value{
A list of cliques. Each clique is a list of Character vertex labels. Ordered by size.
}
\description{
\code{cliqueR.maximal} returns a list of cliques, ordered by size.
}
\details{
A clique is a set of vertices such that every vertex is connected to every other
vertex in the set. Maximal clique returns a list of all the cliques in a graph, from 
3 to the number of vertices in the graph, unless the user specifies otherwise.
}
