% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliqueR.R
\name{cliqueR.maximum}
\alias{cliqueR.maximum}
\title{Maximum clique finder}
\usage{
cliqueR.maximum(filename, sizeOnly = FALSE)
}
\arguments{
\item{filename}{Character path to an edglist graph file.}

\item{sizeOnly}{Logical option to return the size of the maximum clique.
instead of the maximum clique itself.}
}
\value{
A list of character vertex labels in the maximum clique. If sizeOnly is specified, the size of the maximum
  clique is returned.
}
\description{
\code{cliqueR.maximum} returns the maximum clique of a graph file.
}
\details{
A clique is a set of vertices such that every vertex is connected to every other
vertex in the set. The maximum clique of a graph is the clique that has the largest
number of vertices.
}
\examples{
cliqueR.maximum(system.file("extdata","bio-yeast.el",package = "cliqueR"))
cliqueR.maximum(system.file("extdata","bio-yeast.el",package="cliqueR"), TRUE)
}
