% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliqueR.R
\name{cliqueR.paraclique}
\alias{cliqueR.paraclique}
\title{Paraclique enumeration}
\usage{
cliqueR.paraclique(filename, igf = 1, min_mc_size = 5, min_pc_size = 5,
  max_num_pcs = -1)
}
\arguments{
\item{filename}{Character path to a edgelist graph file.}

\item{igf}{Numeric glom factor, 1.0 being the strictest, 0.0 the loosest.}

\item{min_mc_size}{Integer the minimum clique size to include.}

\item{min_pc_size}{Integer the minimum paraclique size to include.}

\item{max_num_pcs}{Integer max number of paracliques to return, default is number of vertices.}
}
\value{
A List of paracliques. Each paraclique itself is a list of Character vertex labels. 
  Ordered by size.
}
\description{
\code{cliqueR.paraclique} returns a list of paracliques, ordered by size.
}
\details{
A clique is a set of vertices such that every vertex is connected to every other
vertex in the set. A paraclique is a clique that includes vertices which are not 
connected to every other vertex in the clique, but are connected to some. The proportion of edges
a vertex has to the size of the core clique is called the glom factor. The glom factor
determines membership of a given vertex to a paraclique. Note that a glom factor approaching
0 will produce numerous large cliques, and significantly increase the risk of the function
never returning for large graphs. 

For questions about algorithms and implementations of the paraclique algorithm, we point users to the papers listed in References.
}
\examples{
cliqueR.paraclique(system.file("extdata","bio-yeast.el",package="cliqueR"))
\donttest{cliqueR.paraclique(system.file("extdata","bio-yeast.el",package="cliqueR"), 0.7)
cliqueR.paraclique(system.file("extdata","bio-yeast.el",package="cliqueR"), 0.7, 6)
cliqueR.paraclique(system.file("extdata","bio-yeast.el",package="cliqueR"), 0.7, 6, 7)
cliqueR.paraclique(system.file("extdata","bio-yeast.el",package="cliqueR"), 0.7, 6, 7, 20)}
}

\references{Chesler EJ, Langston MA. Combinatorial Genetic Regulatory Network Analysis Tools for High Throughput Transcriptomic Data. In: Eskin E, editor. Systems Biology and Regulatory Genomics: Springer; 2006. p. 150–65.

Hagan RD, Langston MA, Wang K. Lower Bounds on Paraclique Density. Discrete Applied Mathematics. 2016;204:208-12.}