% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliqueR.R
\name{formatter}
\alias{formatter}
\title{Graph format checker}
\usage{
formatter(filename)
}
\arguments{
\item{filename}{Character path to an edgelist (.el) file}
}
\value{
If the graph is formatted correctly 0, otherwise -1 and an error
  message is printed.
}
\description{
\code{formatter} verifies that a graph file is correctly formatted for use with cliqueR.
}
\details{
The edgelist must be tab seperated, with the two vertices comprising an edge per line.
the first line of the file must be the number of vertices, tab, 
and the number of edges. If an edge is repeated, there is an edge 
between a vertex and itself, or inconsistencies with the provided
and derived number of vertices or edges, an error will be thrown.
}
