% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_metrics.R
\name{quartet_distance}
\alias{quartet_distance}
\title{Quartet distance between two trees}
\usage{
quartet_distance(tree1, tree2)
}
\arguments{
\item{tree1, tree2}{\code{ape::phylo} objects.}
}
\value{
This development release errors because the metric is not yet
implemented.
}
\description{
This is a reserved API in the current development release and currently
errors.
}
